fix vs 2019 math error

+++ src/operator/bilinear_sampler.cu
@@ -55,8 +55,8 @@ __global__ void BilinearSamplerForwardKernel(const int i_c, const int i_h,
     int grid_index = n * o_h * o_w * 2 + h * o_w + w;
     DType y_real = (*(grid + grid_index + o_h * o_w) + 1) * (i_h - 1) / 2;
     DType x_real = (*(grid + grid_index) + 1) * (i_w - 1) / 2;
-    int top_left_y = static_cast<int>(floor(y_real));
-    int top_left_x = static_cast<int>(floor(x_real));
+    int top_left_y = static_cast<int>(floorf(y_real));
+    int top_left_x = static_cast<int>(floorf(x_real));
     DType top_left_y_w = 1.0 - (y_real - top_left_y);
     DType top_left_x_w = 1.0 - (x_real - top_left_x);
     int data_index = n * i_c * i_h * i_w + c * i_h * i_w + top_left_y * i_w + top_left_x;
@@ -100,8 +100,8 @@ __global__ void BilinearSamplerBackwardKernel(const int i_c, const int i_h,
     DType y_real = (*(grid_src + grid_src_index + o_h * o_w) + 1) * (i_h - 1) / 2;
     DType x_real = (*(grid_src + grid_src_index) + 1) * (i_w - 1) / 2;
 
-    int top_left_y = static_cast<int>(floor(y_real));
-    int top_left_x = static_cast<int>(floor(x_real));
+    int top_left_y = static_cast<int>(floorf(y_real));
+    int top_left_x = static_cast<int>(floorf(x_real));
     DType top_left_y_w = 1.0 - (y_real - top_left_y);
     DType top_left_x_w = 1.0 - (x_real - top_left_x);
     for (int c = 0; c < o_c; ++c) {
