+++ src/operator/spatial_transformer.cu
@@ -64,8 +64,8 @@ BilinearSamplingForwardKernel(const int i_c, const int i_h,
     index_t grid_index = n * o_h * o_w * 2 + h * o_w + w;
     DType y_real = (*(grid + grid_index + o_h * o_w) + 1) * (i_h - 1) / 2;
     DType x_real = (*(grid + grid_index) + 1) * (i_w - 1) / 2;
-    int top_left_y = static_cast<int>(floor(y_real));
-    int top_left_x = static_cast<int>(floor(x_real));
+    int top_left_y = static_cast<int>(floorf(y_real));
+    int top_left_x = static_cast<int>(floorf(x_real));
     DType top_left_y_w = 1.0 - (y_real - top_left_y);
     DType top_left_x_w = 1.0 - (x_real - top_left_x);
     int data_index = n * i_c * i_h * i_w + c * i_h * i_w + top_left_y * i_w + top_left_x;
@@ -117,8 +117,8 @@ BilinearSamplingBackwardKernel(const int i_c, const int i_h,
     index_t grid_src_index = n * o_h * o_w * 2 + h * o_w + w;
     DType y_real = (*(grid_src + grid_src_index + o_h * o_w) + 1) * (i_h - 1) / 2;
     DType x_real = (*(grid_src + grid_src_index) + 1) * (i_w - 1) / 2;
-    int top_left_y = static_cast<int>(floor(y_real));
-    int top_left_x = static_cast<int>(floor(x_real));
+    int top_left_y = static_cast<int>(floorf(y_real));
+    int top_left_x = static_cast<int>(floorf(x_real));
     DType top_left_y_w = 1.0 - (y_real - top_left_y);
     DType top_left_x_w = 1.0 - (x_real - top_left_x);
     for (index_t c = 0; c < o_c; ++c) {
